% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tileserver.R
\name{calculate_tiles}
\alias{calculate_tiles}
\title{Calculate tiles for bbox and zoom}
\usage{
calculate_tiles(bbox, zoom, zoom_guess = 1, min_tiles = c(2, 2),
  messaging = F)
}
\arguments{
\item{bbox}{A bounding box to calculate tiles for}

\item{zoom}{An optional zoom level to calculate tiles for}

\item{zoom_guess}{A starting guess for zoom level needed}

\item{min_tiles}{A minimum number of tiles needed either in total (length 1)
or x and y (length 2, x and y respectively)}
}
\description{
Calculates tiles to span bbox on level zoom.
If zoom is not provided tries to guess zoom level by computing the tiles
until the min_tiles criterion is met.
min_tiles can be length one or two, defining either total tile count, or
dimensions in x and y
}
