% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{crime}
\alias{crime}
\title{Crime data}
\description{
Lightly cleaned Houston crime from January 2010 to August 2010 geocoded with Google Maps
}
\author{
Houston Police Department, City of Houston
}
\references{
\url{http://www.houstontx.gov/police/cs/stats2.htm}
}

