% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tileserver.R
\name{crop_map}
\alias{crop_map}
\title{Crops the map}
\usage{
crop_map(map, bbox, tiles)
}
\arguments{
\item{map}{A ggmap}

\item{bbox}{A bbox}

\item{bbox}{list of tiles in map}
}
\description{
Crop `map` to bbox.
}
\details{
This will at some point be a function usable outside of get_tilemap. For
the time being, don't use this.
}
