% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fetch_region}
\alias{fetch_region}
\title{Fetch tile region.}
\usage{
fetch_region(lon, lat, provider, ..., cache = TRUE, zoom = 10)
}
\description{
Given bounding box, find tiles that span region and stitch them together
into single raster.
}
\examples{
houston <- fetch_region(c(-95.80204, -94.92313), c(29.38048, 30.14344),
  "stamen", "terrain")
ggmap(houston)
}
\keyword{internal}

