% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tileserver.R
\name{format_uri}
\alias{format_uri}
\title{Format URI for tileservers}
\usage{
format_uri(uri, x, y, z)
}
\arguments{
\item{uri}{URI pattern}

\item{x}{x coordinate}

\item{y}{y coordinate}

\item{z}{z coordinate}
}
\description{
Formats a URI for a tileserver. The URI is specified as
`http://example.org/{z}/{x}/{y}.png` or similar.
}
