% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{geocode}
\alias{geocode}
\alias{geocodeQueryCheck}
\title{Geocode}
\usage{
geocode(location, output = c("latlon", "latlona", "more", "all"),
  source = c("google", "dsk"), messaging = FALSE, force = ifelse(source ==
  "dsk", FALSE, TRUE), urlonly = FALSE, override_limit = FALSE,
  nameType = c("long", "short"), ext = "com", inject = "", ...)

geocodeQueryCheck()
}
\arguments{
\item{location}{a character vector of street addresses or place names (e.g.
"1600 pennsylvania avenue, washington dc" or "Baylor University")}

\item{output}{amount of output, "latlon", "latlona", "more", or "all"}

\item{source}{"google" for Google, "dsk" for Data Science Toolkit terminated
its service}

\item{messaging}{turn messaging on/off}

\item{force}{force online query, even if previously downloaded}

\item{urlonly}{return only the url?}

\item{override_limit}{override the current query count
(.GoogleGeocodeQueryCount)}

\item{nameType}{in some cases, Google returns both a long name and a short
name. this parameter allows the user to specify which to grab.}

\item{ext}{top level domain (e.g. "com", "co.nz"); helpful for non-US users}

\item{inject}{character string to add to the url or named character vector of
key-value pairs to be injected (e.g. c("a" = "b") get converted to "a=b"
and appended to the query)}

\item{...}{...}
}
\value{
If \code{output} is "latlon", "latlona", or "more", a data frame. If
  all, a list.
}
\description{
Geocodes a location (find latitude and longitude) using either (1) the Data
Science Toolkit (\url{http://www.datasciencetoolkit.org/about}) or (2) Google
Maps. Note that when using Google you are agreeing to the Google Maps API
Terms of Service at \url{https://developers.google.com/maps/terms}.
}
\details{
Note that the Google Maps API limits to 2500 queries a day. Use
\code{geocodeQueryCheck} to determine how many queries remain.
}
\examples{

\dontrun{ # Server response can be slow; this cuts down check time.

##### basic usage
########################################

geocode("houston texas")

geocode("houston texas", source = "dsk")
geocode("1600 pennsylvania avenue, washington dc")
geocode("the white house")
geocode(c("the white house", "washington dc"))
# see also mutate_geocode()
geocode("ninos", inject = "region=es", urlonly = TRUE)
geocode("ninos", inject = c("region" = "es"), urlonly = TRUE)


##### types of output
########################################

# types of output
geocode("houston texas", output = "latlona")
geocode("houston texas", output = "more")
geocode("Baylor University", output = "more")
str(geocode("Baylor University", output = "all"))


##### interfacing with the google geocoding api
########################################

register_google(key = "your code here")
geocode("houston texas")


# see how many requests we have left with google
geocodeQueryCheck()
geocode("one bear place, waco, texas")
geocode("houston texas", force = TRUE)



##### known issues
########################################
# sources sometime have a hard time reliably geocoding colloquial place names


}

}
\seealso{
\code{\link{mutate_geocode}},
  \url{http://code.google.com/apis/maps/documentation/geocoding/},
  \url{https://developers.google.com/maps/documentation/geocoding/usage-limits}
}
\author{
David Kahle \email{david.kahle@gmail.com}
}
