% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tileserver.R
\name{get_mapnikgrayscale}
\alias{get_mapnikgrayscale}
\title{Get mapnik grayscale maps.}
\usage{
get_mapnikgrayscale(bbox = c(left = -95.80204, bottom = 29.38048, right =
  -94.92313, top = 30.14344), zoom = 10, ...)
}
\arguments{
\item{bbox}{Bounding Box for map}

\item{zoom}{Zoom level of map}

\item{...}{Further arguments passed to get_tilemap}
}
\description{
Gets a map from mapnik (grayscale) defined by bbox and zoom.
Essentially a wrapper around `get_tilemap`.
}
\seealso{
get_tilemap
}
