% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_openstreetmap.R
\name{get_openstreetmap}
\alias{get_openstreetmap}
\title{Get an OpenStreetMap}
\usage{
get_openstreetmap(bbox = c(left = -95.80204, bottom = 29.38048, right =
  -94.92313, top = 30.14344), zoom = 10, crop = TRUE, format = c("png",
  "jpeg", "svg", "pdf", "ps"), messaging = FALSE, urlonly = FALSE,
  color = c("color", "bw"), force = FALSE, ...)
}
\arguments{
\item{bbox}{a bounding box in the format c(lowerleftlon, lowerleftlat,
upperrightlon, upperrightlat)}

\item{zoom}{a zoom level}

\item{crop}{crop raw map tiles to specified bounding box}

\item{format}{character string providing image format - png, jpeg, svg, pdf,
and ps formats}

\item{messaging}{turn messaging on/off}

\item{urlonly}{return url only}

\item{color}{color or black-and-white}

\item{force}{if the map is on file, should a new map be looked
up? Use responsibly.}

\item{...}{...}
}
\value{
a ggmap object (a classed raster object with a bounding box attribute)
}
\description{
\code{get_openstreetmap} accesses a tile server for OpenStreetMap and
downloads/stitches map tiles/formats a map image.  If you don't know how to
get the map you want, go to \url{http://www.openstreetmap.org/}, navigate to
the map extent that you want, click the export tab at the top of the page,
and copy the information into this function.
}
\details{
See \url{http://www.openstreetmap.org/copyright} for license and copyright
information.

As a prerequisite for downloading OpenStreetMap tiles with this function,
persistent tile caching must be enabled by setting the
\code{"ggmap.file_drawer"} \link[=options]{option}.
See \code{\link{file_drawer}}.
}
\examples{

get_openstreetmap(urlonly = TRUE)
\dontrun{
# osm servers get overloaded, which can result in
# erroneous failed checks

# "ggmap.file_drawer" option must be set
osm <- get_openstreetmap()
ggmap(osm)

}

}
\seealso{
\url{http://www.openstreetmap.org/}, \code{\link{ggmap}}
}
\author{
David Kahle \email{david.kahle@gmail.com}
}
