% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pdokmap.r
\name{get_pdokmap}
\alias{get_pdokmap}
\alias{get_pdok_tile_download_fail_log}
\alias{retry_pdok_map_download}
\title{Get a Dutch basemap from PDOK}
\usage{
get_pdokmap(
  bbox = c(left = 5, bottom = 52, right = 5.4, top = 52.2),
  zoom = 10,
  maptype = c("brtachtergrondkaart", "brtachtergrondkaartgrijs",
    "brtachtergrondkaartpastel", "brtachtergrondkaartwater"),
  crop = TRUE,
  messaging = FALSE,
  urlonly = FALSE,
  color = c("color", "bw"),
  force = FALSE,
  where = tempdir(),
  https = TRUE,
  ...
)

get_pdok_tile_download_fail_log()

retry_pdok_map_download()
}
\arguments{
\item{bbox}{a bounding box in the format c(lowerleftlon, lowerleftlat,
upperrightlon, upperrightlat).}

\item{zoom}{a zoom level}

\item{maptype}{brtachtergrondkaart, brtachtergrondkaartgrijs,
brtachtergrondkaartpastel, brtachtergrondkaartwater.}

\item{crop}{crop raw map tiles to specified bounding box. if FALSE, the
resulting map will more than cover the bounding box specified.}

\item{messaging}{turn messaging on/off}

\item{urlonly}{return url only}

\item{color}{color or black-and-white (use force = TRUE if you've already
downloaded the images)}

\item{force}{if the map is on file, should a new map be looked up?}

\item{where}{where should the file drawer be located (without terminating
"/")}

\item{https}{if TRUE, queries an https endpoint so that web traffic between
you and the tile server is ecrypted using SSL.}

\item{...}{...}
}
\value{
a ggmap object (a classed raster object with a bounding box
attribute)
}
\description{
\code{get_pdokmap} accesses a tile server for Dutch maps and
downloads/stitches map tiles/formats a map image. Note that it does not
cover the entire world.
}
\examples{

\dontrun{ some requires Google API key, see ?register_google; heavy network/time load


## basic usage
########################################

bbox <- c(left = 5.0, bottom = 52.0, right = 5.4, top = 52.2)

ggmap(get_pdokmap(bbox, zoom = 13))
ggmap(get_pdokmap(bbox, zoom = 14))
ggmap(get_pdokmap(bbox, zoom = 15))
ggmap(get_pdokmap(bbox, zoom = 16, messaging = TRUE))

place <- "Utrecht"
(google <- get_googlemap(place, zoom = 9))
ggmap(google)
bbox_utrecht <- c(left = 86.05, bottom = 27.21, right = 87.81, top = 28.76)
ggmap(get_pdokmap(bbox_utrecht, zoom = 9))



## map types
########################################

place <- "Amsterdam"
google <- get_googlemap(place, zoom = 10)
ggmap(google)

bbox <- bb2bbox(attr(google, "bb"))

get_pdokmap(bbox, maptype = "brtachtergrondkaart")           \%>\% ggmap()
get_pdokmap(bbox, maptype = "brtachtergrondkaartgrijs")      \%>\% ggmap()
get_pdokmap(bbox, maptype = "brtachtergrondkaartpastel")     \%>\% ggmap()
get_pdokmap(bbox, maptype = "brtachtergrondkaartwater")      \%>\% ggmap()

## zoom levels
########################################

get_pdokmap(bbox, maptype = "brtachtergrondkaartgrijs", zoom = 11) \%>\% ggmap(extent = "device")
get_pdokmap(bbox, maptype = "brtachtergrondkaartgrijs", zoom = 12) \%>\% ggmap(extent = "device")
get_pdokmap(bbox, maptype = "brtachtergrondkaartgrijs", zoom = 13) \%>\% ggmap(extent = "device")
# get_pdokmap(bbox, maptype = "brtachtergrondkaartgrijs", zoom = 14) \%>\% ggmap(extent = "device")
# get_pdokmap(bbox, maptype = "brtachtergrondkaartgrijs", zoom = 15) \%>\% ggmap(extent = "device")
# get_pdokmap(bbox, maptype = "brtachtergrondkaartgrijs", zoom = 16) \%>\% ggmap(extent = "device")
# get_pdokmap(bbox, maptype = "brtachtergrondkaartgrijs", zoom = 17) \%>\% ggmap(extent = "device")
# get_pdokmap(bbox, maptype = "brtachtergrondkaartgrijs", zoom = 18) \%>\% ggmap(extent = "device")


## https
########################################

bbox <- c(left = 5.0, bottom = 52.0, right = 5.4, top = 52.2)
get_pdokmap(bbox, zoom = 14, urlonly = TRUE)
get_pdokmap(bbox, zoom = 14, urlonly = TRUE, https = TRUE)
ggmap(get_pdokmap(bbox, zoom = 15, https = TRUE, messaging = TRUE))



## known issues
########################################

# in some cases PDOK/NGR servers will not return a tile for a given map
# this tends to happen in high-zoom situations, but it is not always
# clear why it happens. these tiles will appear as blank parts of the map.

# ggmap provides some tools to try to recover the missing tiles, but the
# servers seem pretty persistent at not providing the maps.

bbox <- c(left = 5.0, bottom = 52.0, right = 5.4, top = 52.2)
ggmap(get_pdokmap(bbox, zoom = 17))
get_pdok_tile_download_fail_log()
retry_pdok_map_download()




}


}
\seealso{
\url{https://www.pdok.nl/introductie/-/article/basisregistratie-topografie-achtergrondkaarten-brt-a-}, \code{\link[=ggmap]{ggmap()}}
}
