% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tileserver.R
\name{get_tile}
\alias{get_tile}
\title{Gets a tile}
\usage{
get_tile(x, y, z, uri_patterns, max_tries = length(uri_patterns) * 3,
  start_index = 1, messaging = F, force = F)
}
\arguments{
\item{x}{X coordinate of tile}

\item{y}{Y coordinate of tile}

\item{z}{Z coordinate of tile}

\item{uri_patterns}{Patterns describing one or several tile servers where
the tiles can be found. Should be formatted as described in `format_uri`.}

\item{max_tries}{Maximum tries until failure}

\item{start_index}{Index of uri_patterns to start getting data from. On
failure is incremented to try the next server: i=i+1%%length(uri_patterns)+1}

\item{messaging}{print messages in console}

\item{force}{ignore cached requests}
}
\value{
a ggmap object with the tile that was requested
}
\description{
Tries to get tile `x`, `y`, `z` from tile server(s) defined in
`uri_patterns`. On failure recurses with a decreased max_tries until
max_tries reaches zero.
}
