\name{GetMap.OSM}
\alias{GetMap.OSM}
\title{Grab an OSM map.}
\usage{
  GetMap.OSM(lonR = c(-74.02132, -73.98622),
    latR = c(40.69983, 40.72595), scale = 20000,
    destfile = "MyTile.png", format = "png",
    RETURNIMAGE = TRUE, GRAYSCALE = FALSE, NEWMAP = TRUE,
    verbose = 1, ...)
}
\arguments{
  \item{lonR}{longitude range}

  \item{latR}{latitude range}

  \item{scale}{see \code{\link{GetMap.OSM}} in package
  RgoogleMaps}

  \item{destfile}{see \code{\link{GetMap.OSM}} in package
  RgoogleMaps}

  \item{format}{see \code{\link{GetMap.OSM}} in package
  RgoogleMaps}

  \item{RETURNIMAGE}{see \code{\link{GetMap.OSM}} in
  package RgoogleMaps}

  \item{GRAYSCALE}{see \code{\link{GetMap.OSM}} in package
  RgoogleMaps}

  \item{NEWMAP}{see \code{\link{GetMap.OSM}} in package
  RgoogleMaps}

  \item{verbose}{see \code{\link{GetMap.OSM}} in package
  RgoogleMaps}

  \item{...}{see \code{\link{GetMap.OSM}} in package
  RgoogleMaps}
}
\value{
  see \code{\link{GetMap.OSM}} in package RgoogleMaps
}
\description{
  This is virtually identical to the function
  \code{\link{GetMap.OSM}} in the RgoogleMaps package.
}
\author{
  Markus Loecher, Sense Networks
  \email{markus@sensenetworks.com}
}
\seealso{
  \code{\link{GetMap.OSM}} in package RgoogleMaps
}

