% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMapTiles.R
\name{GetMapTiles}
\alias{GetMapTiles}
\title{Download Map Tiles to local directory}
\usage{
GetMapTiles(bbox = c(left = -95.80204, bottom = 29.38048, right = -94.92313,
  top = 30.14344), zoom = 10, urlBase = "http://tile.stamen.com/terrain/",
  CheckExistingFiles = TRUE, TotalSleep = NULL, tileExt = ".png",
  tileDir = "~/mapTiles/stamen/", verbose = 0)
}
\arguments{
\item{bbox}{Extent over which to pull tiles}

\item{zoom}{map zoom, an integer from 3 (continent) to 21
(building), default value 10 (city).  openstreetmaps limits a
zoom of 18, and the limit on stamen maps depends on the
maptype.  "auto" automatically determines the zoom for bounding
box specifications, and is defaulted to 10 with center/zoom
specifications.  maps of the whole world currently not
supported.}

\item{urlBase}{path of tile server to query}

\item{CheckExistingFiles}{will check if current tiles are already downloaded and skip if are}

\item{TotalSleep}{##<< overall time (in seconds) that one is willing to add in
between downloads. This is intended to lower the risk of a server denial. 
If NULL no call to \link{Sys.sleep} is executed}

\item{tileExt}{file extension of files from tile server}

\item{tileDir}{Where are local tiles stored.  Only relevant if doOffline == TRUE}

\item{verbose}{level of verbosity}
}
\value{
a ggmap object (a classed raster object with a bounding
  box attribute)
}
\description{
\code{GetMapTiles} is a smart wrapper that queries the Google Maps,
OpenStreetMap, Stamen Maps or Naver Map servers for a map.  This download map
 tiles from specified map tile servers such as openstreetmap or Google
 Query the server for map tiles, defined uniquely by their X and Y ID and zoom.
 For offline usage, these map tiles are stored in a local directory
 THIS CODE WAS MODIFIED FROM THE FUNCTION IN THE RgoogleMaps Package
}
\examples{
GetMapTiles(zoom=9
                      ,bbox = sbbox
                      ,verbose=1
                      ,urlBase = "http://tile.stamen.com/terrain/"
                      ,tileDir= "./mapTiles/stamen/terrain/")
}
\seealso{
\code{\link{ggmap}}, \code{\link{GetMap}} in package
  RgoogleMaps
}
\author{
David Kahle \email{david.kahle@gmail.com}
}
