% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmap-defunct.R
\name{ggmap-defunct}
\alias{ggmap-defunct}
\title{Defunct ggmap functions}
\description{
As provider services change over time, ggmap has to make corresponding
changes.  Since its inception, a few services have stopped offering their
previous functionality, and in some cases this has required us to remove
those functions from the package entirely.
}
\details{
The following are defunct ggmap functions:
\itemize{
\item get_cloudemademap
\item get_navermap
\item get_openstreetmap
}
}
