% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmap-package.R
\docType{package}
\name{ggmap-package}
\alias{ggmap-package}
\alias{ggmappackage}
\title{Download and Process GIMMS3g Data}
\description{
Spatial Visualization with ggplot2
}
\author{
David Kahle
}
\keyword{package}

