% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmap2raster.R
\name{ggmap2raster}
\alias{ggmap2raster}
\title{Rasterize a 'ggmap'}
\usage{
ggmap2raster(map, crs = NULL)
}
\arguments{
\item{map}{A 'ggmap' object.}

\item{crs}{Optional output coordinate reference system (CRS) passed on to 
\code{\link{projectRaster}}.}
}
\value{
A red-green-blue 'RasterStack' object.
}
\description{
\code{ggmap2raster} converts a 'ggmap' object (created via 
\code{\link{get_map}}) to a multi-layered (i.e., red-green-blue) 
'RasterStack' object.
}
\examples{
## sample data
map_world <- get_map(location = "world")

## convert 'ggmap' to 'RasterStack'
ggmap2raster(map_world)

}
\author{
Florian Detsch (\email{florian.detsch@staff.uni-marburg.de})
}
\seealso{
\code{\link{get_map}}, \code{\link{projectRaster}}.
}

