% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmap_credentials.R
\name{ggmap_credentials}
\alias{ggmap_credentials}
\alias{print.ggmap_credentials}
\title{ggmap Credentials}
\usage{
ggmap_credentials()

\method{print}{ggmap_credentials}(x, ...)
}
\arguments{
\item{x}{the ggmap credentials object; getOption("ggmap")}

\item{...}{... (dumped)}
}
\description{
Get ggmap's credentials; these are all Google credentials at this
point.
}
\examples{

ggmap_credentials()

}
\seealso{
\code{\link{register_google}}
}
\author{
David Kahle \email{david.kahle@gmail.com}
}
