% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmap_options.R
\name{ggmap_options}
\alias{ggmap_options}
\alias{set_ggmap_option}
\alias{ggmap_credentials}
\title{ggmap Options}
\usage{
set_ggmap_option(...)

ggmap_credentials()
}
\arguments{
\item{...}{a named listing of options to set}
}
\description{
ggmap stores options as a named list in R's global options, i.e.
\code{getOption("ggmap")}. It currently stores two such options, one for
Google credentialing and one to suppress private API information in the URLs
printed to the screen when web queries are placed. For both of those, see
\code{\link{register_google}}.
}
\examples{

getOption("ggmap")

}
\seealso{
\code{\link{register_google}}
}
\author{
David Kahle \email{david.kahle@gmail.com}
}
