% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmap_to_webmercator.R
\name{ggmap_to_webmercator}
\alias{ggmap_to_webmercator}
\title{\code{ggmap_to_webmercator} change bounding box to match with sf objects (CRS needs to be 3857)}
\usage{
ggmap_to_webmercator(map)
}
\arguments{
\item{map}{The map to update}
}
\description{
\code{ggmap_to_webmercator} change bounding box to match with sf objects (CRS needs to be 3857)
}
