% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inset_raster.R
\name{inset_raster}
\alias{inset_ggmap}
\alias{inset_raster}
\title{Create a (ggplot2) raster layer}
\usage{
inset_raster(raster, xmin, xmax, ymin, ymax, interpolate = FALSE)

inset_ggmap(ggmap)
}
\arguments{
\item{raster}{raster object to display}

\item{xmin, xmax}{x location (in data coordinates) giving
horizontal location of raster}

\item{ymin, ymax}{y location (in data coordinates) giving vertical
location of raster}
}
\description{
This is a special version of ggplot2::annotation_raster for use
with ggmap.  (It simply removes the requirement for cartesian
coordinates.)  The only difference between \code{inset_raster}
and \code{inset_ggmap} is their arguments.  \code{inset_ggmap} is
simply a wrapper of \code{inset_raster} with \code{xmin}, ...,
\code{ymax} arguments equal to the map's bounding box.
}
\examples{

# see ?route



}

