% Generated by roxygen2 (4.0.2): do not edit by hand
\name{LonLat2XY}
\alias{LonLat2XY}
\title{Convert a lon/lat coordinate to a tile coordinate}
\usage{
LonLat2XY(lon_deg, lat_deg, zoom, xpix = 256, ypix = 256)
}
\arguments{
\item{lon_deg,}{lat_deg longitude and latitutde in degrees}

\item{zoom}{zoom}

\item{xpix,ypix}{height and width of tile in pixels}
}
\value{
A data frame with columns X, Y (tile), x, y (position within tile)
}
\description{
Convert a lon/lat coordinate to a tile coordinate
}
\examples{
baylor <- list(lon = -97.11844, lat = 31.54822)
LonLat2XY(baylor$lon, baylor$lat, 10)
}
\author{
David Kahle \email{david.kahle@gmail.com}, based on function LatLon2XY by Markus Loecher, Sense Networks \email{markus@sensenetworks.com} in package RgoogleMaps
}
\seealso{
\url{http://wiki.openstreetmap.org/wiki/Slippy_map_tilenames}
}

