% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route.R
\name{route}
\alias{route}
\title{Grab a route from Google}
\usage{
route(from, to, mode = c("driving", "walking", "bicycling", "transit"),
  structure = c("legs", "route"), output = c("simple", "all"),
  alternatives = FALSE, units = "metric", messaging = FALSE,
  override_limit = FALSE, ...)
}
\arguments{
\item{from}{name of origin addresses in a data frame (vector
accepted)}

\item{to}{name of destination addresses in a data frame (vector
accepted)}

\item{mode}{driving, bicycling, walking, or transit}

\item{structure}{structure of output, see examples}

\item{output}{amount of output}

\item{alternatives}{should more than one route be provided?}

\item{units}{"metric"}

\item{messaging}{turn messaging on/off}

\item{override_limit}{override the current query count
(.GoogleRouteQueryCount)}

\item{...}{...}
}
\value{
a data frame (output="simple") or all of the geocoded
  information (output="all")
}
\description{
Grab a route from Google. Note that in most cases by using this
function you are agreeing to the Google Maps API Terms of Service
at https://developers.google.com/maps/terms.
}
\examples{

\dontrun{ # to cut down on check time

from <- "houson, texas"
to <- "waco, texas"
route_df <- route(from, to, structure = "route")
trek_df  <-  trek(from, to, structure = "route")
qmap("college station, texas", zoom = 8) +
  geom_path(
    aes(x = lon, y = lat),  colour = "red",
    size = 1.5, alpha = .5,
    data = route_df, lineend = "round"
  ) +
  geom_path(
    aes(x = lon, y = lat),  colour = "blue",
    size = 1.5, alpha = .5,
    data = trek_df, lineend = "round"
  )

qmap("college station, texas", zoom = 6) +
  geom_path(
    aes(x = lon, y = lat), colour = "red", size = 1.5,
    data = route_df, lineend = "round"
  )

routeQueryCheck()



}

}
\author{
David Kahle \email{david.kahle@gmail.com}
}
\seealso{
\url{https://developers.google.com/maps/documentation/directions/},
\code{\link{trek}}, \code{\link{legs2route}},
\code{\link{routeQueryCheck}}, \code{\link{geom_leg}},
\code{\link{register_google}}
}

