% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_map_alpha.R
\name{set_map_alpha}
\alias{set_map_alpha}
\title{Add transparency to a ggmap object}
\usage{
set_map_alpha(map, alpha)
}
\arguments{
\item{map}{A \code{ggmap} object.}

\item{alpha}{The required transparency (a number between 0 and 1).}
}
\value{
A \code{ggmap} object.
}
\description{
Add an alpha (transparency) layer to a \code{ggmap} object. This makes it
possible to superimpose multiple map layers and have the lower layers shine
through.
}
\examples{
zoom <- 12
bbox <- c(-0.3275, 51.407222, 0.0725, 51.607222)

base <- get_map(bbox, zoom, maptype = "terrain-lines")
overlay <- get_map(bbox, zoom, maptype = "watercolor")

# Plot the base map.
ggmap(base)
# Plot the overlay map (100\% opacity).
ggmap(overlay)

# Plot the base map with the overlay superimposed at 25\% opacity.
ggmap(base) +
  inset_ggmap(
    set_map_alpha(overlay, 0.25)
  )
}
