% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urlsigning.R
\name{signurl}
\alias{signurl}
\title{Sign a url using a google digital signature}
\usage{
signurl(input_url, secret)
}
\arguments{
\item{input_url}{The url to be signed. Should include a client field.}

\item{secret}{The private key}
}
\value{
The signed url
}
\description{
Sign a url using a google digital signature
}
\details{
Derived from the python urlsigner: 
https://raw.githubusercontent.com/googlemaps/js-v2-samples/61b3f58eb1286a428843f8401048226b8648a76b/urlsigning/urlsigner.py
}
\author{
Richard Beare
}

