% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tileserver.R
\name{tile_stitch}
\alias{tile_stitch}
\title{Stitch list of ggmaps}
\usage{
tile_stitch(tiles)
}
\arguments{
\item{tiles}{A list of ggmap objects to stitch}
}
\value{
A single ggmap object stitched from the other ggmap objects.
}
\description{
Stitches a list of ggmap objects, as they are for example provided by
get_tilemap or get_tile.
}
