% Generated by roxygen2 (4.0.1): do not edit by hand
\name{timeStamp}
\alias{timeStamp}
\title{Convenient time grabber}
\usage{
timeStamp()
}
\value{
time as a character string
}
\description{
Convenient time grabber
}
\examples{
\dontrun{

timeStamp()
str(timeStamp())

}
}
\author{
David Kahle \email{david.kahle@gmail.com}.
}

