% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trek.R
\name{trek}
\alias{trek}
\title{Grab a trek from Google}
\usage{
trek(from, to, mode = c("driving", "walking", "bicycling", "transit"),
  structure = c("legs", "route"), output = c("simple", "all"),
  alternatives = FALSE, messaging = FALSE, sensor = FALSE,
  override_limit = FALSE)
}
\arguments{
\item{from}{name of origin addresses in a data frame (vector
accepted)}

\item{to}{name of destination addresses in a data frame (vector
accepted)}

\item{mode}{driving, bicycling, walking, or transit}

\item{structure}{structure of output, see examples}

\item{output}{amount of output}

\item{alternatives}{should more than one route be provided?}

\item{messaging}{turn messaging on/off}

\item{sensor}{whether or not the geocoding request comes from a
device with a location sensor}

\item{override_limit}{override the current query count
(.GoogleRouteQueryCount)}
}
\value{
a data frame (output="simple") or all of the geocoded
  information (output="all")
}
\description{
Grab a trek (curvy route) from Google. Note that in most cases by
using this function you are agreeing to the Google Maps API Terms
of Service at https://developers.google.com/maps/terms.
}
\examples{

\dontrun{ # to cut down on check time

from <- "houson, texas"; to <- "waco, texas"
trek_df <- trek(from, to, structure = "route", alternatives = TRUE)

qmplot(lon, lat, data = trek_df, geom = "path")

qmplot(lon, lat, data = trek_df, geom = "path",
  color = route, size = I(2)
)


from <- "los angeles, california"; to <- "new york, new york"
trek_df <- trek(from, to, structure = "route", alternatives = TRUE)
qmplot(lon, lat, data = trek_df, geom = "path",
  color = route, size = I(2), maptype = "terrain"
) + facet_wrap(~ route, ncol = 1)




}

}
\author{
David Kahle \email{david.kahle@gmail.com} with the key
  decoding algorithm due to stackoverflow user akhmed
}
\seealso{
\url{https://developers.google.com/maps/documentation/directions/},
\url{http://stackoverflow.com/questions/30270011/ggmap-route-finding-doesnt-stay-on-roads},
\code{\link{route}}, \code{\link{routeQueryCheck}}
}

