\docType{data}
\name{zips}
\alias{zips}
\title{Zip code data for the Greater Houston Metropolitan Area from the 2000 census}
\description{
  Zip code data for the Greater Houston Metropolitan Area
  from the 2000 census
}
\author{
  U.S. Census Bureau, Geography Division, Cartographic
  Products Management Branch
}
\references{
  \url{http://www.census.gov/geo/www/cob/z52000.html}
}
\keyword{datasets}

