% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mapdist.R
\name{distQueryCheck}
\alias{distQueryCheck}
\title{Check Google Maps Distance Matrix API query limit}
\usage{
distQueryCheck()
}
\value{
a data frame
}
\description{
Check Google Maps Distance Matrix API query limit
}
\examples{
distQueryCheck()
}
\author{
David Kahle \email{david.kahle@gmail.com}
}
\seealso{
\url{http://code.google.com/apis/maps/documentation/distancematrix/}
}

