\name{geocodeQueryCheck}
\alias{geocodeQueryCheck}
\title{Check Google Geocoding API query limit}
\usage{
  geocodeQueryCheck()
}
\value{
  a data frame
}
\description{
  Check Google Geocoding API query limit
}
\examples{
geocodeQueryCheck()
}
\author{
  David Kahle \email{david.kahle@gmail.com}
}
\seealso{
  \url{https://developers.google.com/maps/documentation/geocoding/}
}

