% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ggmap}
\alias{print.ggmap}
\title{Print a map}
\usage{
\method{print}{ggmap}(x, ...)
}
\arguments{
\item{x}{an object of class elicit}

\item{...}{additional parameters}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a console description of a map
}
\examples{


get_stamenmap(zoom = 9)


}
