% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_google.R
\name{register_google}
\alias{check_google_for_error}
\alias{goog_account}
\alias{goog_client}
\alias{goog_day_limit}
\alias{goog_key}
\alias{goog_second_limit}
\alias{goog_signature}
\alias{has_goog_account}
\alias{has_goog_client}
\alias{has_goog_key}
\alias{has_goog_signature}
\alias{register_google}
\title{Register a Google API}
\usage{
register_google(key, account_type, client, signature, second_limit, day_limit)

goog_key()

has_goog_key()

has_goog_account()

goog_account()

goog_client()

has_goog_client()

goog_signature()

has_goog_signature()

goog_second_limit()

goog_day_limit()

check_google_for_error(tree)
}
\arguments{
\item{key}{an api key}

\item{account_type}{\code{"standard"} or \code{"premium"}}

\item{client}{client code}

\item{signature}{signature code}

\item{second_limit}{query limit per second (default 50)}

\item{day_limit}{query limit per day (default 2500 for standard
accounts, 100000 for premium accounts)}

\item{tree}{a json tree from \code{\link{fromJSON}}}
}
\description{
Many features of the Google web mapping services can be improved
by using standard or premium credentials, such as usage limits
and query rates. This function allows users to input this
information into R as a global option to be retained for the
entire session.
}
\examples{


has_goog_key()
goog_key()
has_goog_client()
has_goog_signature()
\dontrun{
register_google(key = "[your key here]")
}

}
\author{
David Kahle \email{david.kahle@gmail.com}
}
\seealso{
\url{http://code.google.com/apis/maps/documentation/geocoding/},
\url{https://developers.google.com/maps/documentation/geocoding/usage-limits}
}

