% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revgeocode.R
\name{revgeocode}
\alias{revgeocode}
\title{Reverse geocode}
\usage{
revgeocode(location, output = c("address", "more", "all"),
  messaging = FALSE, urlonly = FALSE, override_limit = FALSE,
  ext = "com", inject = "", ...)
}
\arguments{
\item{location}{a location in longitude/latitude format}

\item{output}{amount of output}

\item{messaging}{turn messaging on/off}

\item{urlonly}{return only the url?}

\item{override_limit}{override the current query count
(.GoogleGeocodeQueryCount)}

\item{ext}{domain extension (e.g. "com", "co.nz")}

\item{inject}{character string to add to the url}

\item{...}{...}
}
\value{
depends (at least an address)
}
\description{
reverse geocodes a longitude/latitude location using Google Maps.
Note that in most cases by using this function you are agreeing
to the Google Maps API Terms of Service at
https://developers.google.com/maps/terms.
}
\examples{

\dontrun{ # Server response can be slow; this cuts down check time.

( gc <- as.numeric(geocode("the white house")) )
revgeocode(gc)
revgeocode(gc, output = "more")
revgeocode(gc, output = "all")
geocodeQueryCheck()

}

}
\seealso{
\url{http://code.google.com/apis/maps/documentation/geocoding/}
}
\author{
David Kahle \email{david.kahle@gmail.com}
}
