\name{routeQueryCheck}
\alias{routeQueryCheck}
\title{Check Google Maps Directions API query limit}
\usage{
  routeQueryCheck()
}
\value{
  a data frame
}
\description{
  Check Google Maps Directions API query limit
}
\examples{
\dontrun{
routeQueryCheck()
}
}
\author{
  David Kahle \email{david.kahle@gmail.com}
}
\seealso{
  \url{https://developers.google.com/maps/documentation/directions/}
}

