% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trek.R
\name{trek}
\alias{trek}
\title{Grab a trek from Google}
\usage{
trek(from, to, mode = c("driving", "walking", "bicycling", "transit"),
  output = c("simple", "all"), alternatives = FALSE, units = "metric",
  messaging = FALSE, urlonly = FALSE, override_limit = FALSE,
  ext = "com", inject = "", ...)
}
\arguments{
\item{from}{name of origin addresses in a data frame (vector
accepted)}

\item{to}{name of destination addresses in a data frame (vector
accepted)}

\item{mode}{driving, bicycling, walking, or transit}

\item{output}{amount of output}

\item{alternatives}{should more than one route be provided?}

\item{units}{"metric"}

\item{messaging}{turn messaging on/off}

\item{urlonly}{return only the url?}

\item{override_limit}{override the current query count
(.GoogleRouteQueryCount)}

\item{ext}{domain extension (e.g. "com", "co.nz")}

\item{inject}{character string to add to the url}

\item{...}{...}
}
\value{
a data frame (output="simple") or all of the geocoded
  information (output="all")
}
\description{
Grab a trek from Google that will plot over the roadways. Note
that in most cases by using this function you are agreeing to the
Google Maps API Terms of Service at
https://developers.google.com/maps/terms.
}
\examples{

\dontrun{ # to cut down on check time

from <- "houston, texas"
to <- "waco, texas"
route_df <- route(from, to, structure = "route")
trek_df <- trek(from, to, structure = "route")
qmap("college station, texas", zoom = 8) +
  geom_path(
    aes(x = lon, y = lat),  colour = "red",
    size = 1.5, alpha = .5,
    data = route_df, lineend = "round"
  ) +
  geom_path(
    aes(x = lon, y = lat),  colour = "blue",
    size = 1.5, alpha = .5,
    data = trek_df, lineend = "round"
  )



from <- "rice university houston texas"
to <- "1001 Bissonnet St, Houston, TX 77005"
trek_df <- trek(from, to)
qmplot(lon, lat, data = trek_df, geom = "path", maptype = "terrain",
  color = I("red"), size = I(2), alpha = I(.5)
)

trek_df <- trek(from, to, mode = "walking")
qmplot(lon, lat, data = trek_df, geom = "path", maptype = "terrain",
  color = I("red"), size = I(2), alpha = I(.5)
)

trek_df <- trek(from, to, mode = "transit")
qmplot(lon, lat, data = trek_df, geom = "path", maptype = "terrain",
  color = I("red"), size = I(2), alpha = I(.5)
)


from <- "houston, texas"; to <- "waco, texas"
trek_df <- trek(from, to, alternatives = TRUE)
qmplot(lon, lat, data = trek_df, geom = "path",
  color = route, size = I(2), alpha = I(.5)
)


from <- "los angeles, california"; to <- "new york, new york"
trek_df <- trek(from, to, alternatives = TRUE)
qmplot(lon, lat, data = trek_df, geom = "path",
  color = route, size = I(2), maptype = "terrain",
  zoom = 5
) + facet_wrap(~ route, ncol = 1)




}

}
\seealso{
\url{https://developers.google.com/maps/documentation/directions/},
\url{http://stackoverflow.com/questions/30270011/ggmap-route-finding-doesnt-stay-on-roads},
\code{\link{route}}, \code{\link{routeQueryCheck}}
\code{\link{register_google}}
}
\author{
David Kahle \email{david.kahle@gmail.com} with the key
  decoding algorithm due to stackoverflow user akhmed
}
