% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wind.R
\docType{data}
\name{wind}
\alias{wind}
\title{Wind data from Hurricane Ike}
\description{
Wind data from Hurricane Ike
}
\details{
Powell, M. D., S. H. Houston, L. R. Amat, and N Morisseau-Leroy, 1998: The HRD real-time hurricane wind analysis system. J. Wind Engineer. and Indust. Aerodyn. 77&78, 53-64
}
\author{
Atlantic Oceanographic and Meteorological Laboratory (AOML), a division of the National Oceanic and Atmospheric Administration (NOAA)
}
\references{
\url{http://www.aoml.noaa.gov/hrd/Storm_pages/ike2008/wind.html}
}

