% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zips.R
\docType{data}
\name{zips}
\alias{zips}
\title{Zip code data for the Greater Houston Metropolitan Area from the 2000 census}
\description{
Zip code data for the Greater Houston Metropolitan Area from the 2000 census
}
\author{
U.S. Census Bureau, Geography Division, Cartographic Products Management Branch
}
\references{
Downloaded from http://www.census.gov/geo/www/cob/z52000.html (now defunct).
}

